Move node caching to NodeCacheManager
